app/database/connection.py
import mariadb
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def get_connection():
    try:
        conn = mariadb.connect(
            host=os.getenv('DB_HOST'),
            port=int(os.getenv('DB_PORT')),
            user=os.getenv('DB_USER'),
            password=os.getenv('DB_PASSWORD'),
            database=os.getenv('DB_NAME')
        )
        return conn
    except mariadb.Error as e:
        print(f"Error connecting to MariaDB: {e}")
        return None
