requirements.txt
pytube
python-dotenv
mariadb
beautifulsoup4
requests
yt_dlp